pyproject.toml
[tool.poetry]
name = "mason-server"
version = "0.0.0"
description = "Backend server for mason framework."
authors = []
packages = [
  { include = "mason_server"}
]

[tool.poetry.dependencies]
python = "^3.7"
importlib-metadata = "^1.6.0"
sanic = "^20.3.0"
mason-framework = "^0.0.5"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
tox = "^3.14.6"
coverage = "^5.0.4"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-asyncio = "^0.11.0"
mock = "^4.0.2"
asyncmock = "^0.4.2"
pytest-sugar = "^0.9.3"
aioconsole = "^0.1.16"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
